src/cluster-mgmt/src/system-maintenance/requirements.txt
asyncssh==2.14.2
aiohttp>=3.10.0

# system client
absl-py==2.1.0
grpcio==1.64.1
grpcio-tools==1.64.1
kubernetes>=32.0.1
mypy-protobuf==3.6.0
paramiko==3.4.0
protobuf==5.27.2
sshtunnel==0.4.0
types-protobuf==5.26.0.20240422

# for testing
pytest>=8.2.1
pytest-mock>=3.14.0
pytest-asyncio==0.23.7